move ExecCtxt to executor and directly return it
